+++ unit_tests/main.cpp
@@ -36,7 +36,10 @@ bool verboseMode = false;
 GTEST_API_ int main(int argc, char **argv) {
 	testing::InitGoogleTest(&argc, argv);
 	// uncomment if you only want to run selected tests
-	//::testing::GTEST_FLAG(filter) = "*testFasterEngineSpinningUp*";
+	/**
+	 * See TEST_FROM_TRIGGER_ID to limit test just for last trigger
+	 */
+	//::testing::GTEST_FLAG(filter) = "*AllTriggersFixture*";
 	int result = RUN_ALL_TESTS();
 	// windows ERRORLEVEL in Jenkins batch file seems to want negative value to detect failure
 	return result == 0 ? 0 : -1;