+++ firmware/CHANGELOG.md
@@ -30,6 +30,7 @@ All notable user-facing or behavior-altering changes will be documented in this 
 ### Added
  - GM 60/2/2/2 trigger https://github.com/rusefi/rusefi/wiki/All-Supported-Triggers#gm
  - TriTach trigger https://github.com/rusefi/rusefi/wiki/All-Supported-Triggers#tritach
+ - Skoda Favorit trigger https://github.com/rusefi/rusefi/wiki/All-Supported-Triggers#skoda-favorit
 
 ### 2021 Printing Ink Day